requirements.txt

aws-cdk.assets==1.72.0
aws-cdk.aws-apigateway==1.72.0
aws-cdk.aws-applicationautoscaling==1.72.0
aws-cdk.aws-autoscaling-common==1.72.0
aws-cdk.aws-certificatemanager==1.72.0
aws-cdk.aws-cloudformation==1.72.0
aws-cdk.aws-cloudwatch==1.72.0
aws-cdk.aws-codeguruprofiler==1.72.0
aws-cdk.aws-ec2==1.72.0
aws-cdk.aws-efs==1.72.0
aws-cdk.aws-elasticloadbalancingv2==1.72.0
aws-cdk.aws-events==1.72.0
aws-cdk.aws-iam==1.72.0
aws-cdk.aws-kms==1.72.0
aws-cdk.aws-lambda==1.72.0
aws-cdk.aws-logs==1.72.0
aws-cdk.aws-rds==1.72.0
aws-cdk.aws-route53==1.72.0
aws-cdk.aws-s3==1.72.0
aws-cdk.aws-s3-assets==1.72.0
aws-cdk.aws-sam==1.72.0
aws-cdk.aws-secretsmanager==1.72.0
aws-cdk.aws-ssm==1.72.0
aws-cdk.cloud-assembly-schema==1.72.0
aws-cdk.core==1.72.0
aws-cdk.custom-resources==1.72.0
aws-cdk.cx-api==1.72.0
aws-cdk.region-info==1.72.0
boto3==1.16.16
botocore==1.19.16

